Modify path (-p) handling
a)
The path should be handled like a regular shell path variable,
i.e. using the colon as a separator.
b)
When there is only one path it should be wrapped in a one-element
list, otherwise an error condition will occur in _executables() in the
list comprehension (which will iterate over the option's *characters*)

But actually the path handling should be changed generally.
I think (if I'm not mistaken) the paths given should be *base*
paths from which all child directories with lilypond installations
can be inferred (both binary releases and local builds, with their
differing directory structures).