Properly handle relative paths (bug fix)
When input files are entered using relative paths (i.e. not either
in the current dir or specified as absolute paths) all the functions
failed.

This is treated by determining the absolute path of the input files
and composing all relevant file names as absolute paths.

(Change handling of file extensions as well)