Properly handle LilyPond failures (no score created)
If one or both LilyPond versions fail to produce a score
an informative error message is printed (instead of the obscure
FileNotFoundError) and the script is aborted.

Note that this is based on the previous change to purge old files
before compilation.

Note also that we can't rely on LilyPond's exit code because LilyPond
already "fails" when the \version is too old.